third_party/WebKit/LayoutTests/http/tests/css/resource-timing-details-for-revalidation-expected.txt

CONSOLE WARNING: document.fonts.ready() method is going to be replaced with document.fonts.ready attribute in future releases. Please be prepared. For more help, check https://code.google.com/p/chromium/issues/detail?id=392077#c3 .
CONSOLE WARNING: document.fonts.ready() method is going to be replaced with document.fonts.ready attribute in future releases. Please be prepared. For more help, check https://code.google.com/p/chromium/issues/detail?id=392077#c3 .
PASS successfullyParsed is true

TEST COMPLETE
PASS entry.domainLookupStart > 0 is true
This fetches ahem font.
